Refactor orphaned records handling
The `orphan` attribute is renamed to `orphanRemoval` with
inverse meaning and added support for this attribute on
one-to-one field too.

The default behavior on one-to-many is to remove orphaned
records if o2m field is backed by an inverse field (mappedBy)
but keep orphaned records if o2m field is not mapped.

The default behavior on one-to-one field is to keep orphaned
records.

Old domain definitions should me migrated by replacing `orphan="true"`
with `orphanRemoval="false"`.