Fix incompatible type issue in action-group
When any of the action in the group returns a map with key
"value" the action-group tries to update context with that
value so that next actions can utilise this updated value
from context.

However, if the value is not compatible to the target field,
it may throw IllegalArgumentException.

This change will try to automatically convert the incompatible
type with type adapters.

Fixes RM-5373