Add support for seamless access to custom fields
If the 'json=true' field be accessed with '$' prefix, a special
proxy context is returned, which can be used to seamlessly access
custom fields associated with the field.

Also, if access to a field doesn't resolve to any real field, and
a custom field with same name exists on "attrs" json field, it will
be automatically resolved as custom field.