Add 'depends' attribute on form widgets
The depends attribute can be used to specify comma separated list
of field names the widget depends on (for js expressions or any
other way).

If depended fields are not included in form view, it will be added
to the list of fields to fetch.

In future, we'll use this information to optimize js expression
evaluation.

See RM-3911