Fix bulk update/delete issue with MySQL
The bulk update/delete operations with filters can't work
on MySQL as it doesn't allow sub-select on same table and
JPQL doesn't support JOIN on update/delete queries.

The bulk update/delete api is re-worked to check if mysql
is used and in that case, will do update/delete in batches.

Fixes RM-12352