Support x-field attribute with InfoButton widget
Use the `x-field` attribute on `info-button` widget to specify the
bound field. When using `x-field`, the button and the field are 2
distinct elements. Any attributes defined on that field will be used to
format the value. Moreover, this allows to change the button attributes
without impact on the bound field.

RM-43369