Fix duplicate publication warning
When `java-gradle-plugin` is used, refer to `pluginMaven` publication
instead of the `mavenJava`.

Also, there is no need to redeclare from `components.java`, which was
already done by the plugin.

This fixes following warning : `Multiple publications with coordinates
'...' are published to repository '...'. The publications will
overwrite each other!` when running `publishToMavenLocal`.

Signed-off-by: Pierre Belloy <p.belloy@axelor.com>