Fix form layout issue
To keep the form layout as per old implementation, we can't always
hide grid items that doesn't fill the layout. It is causing extra
spacing if all items in the row are hidden.

This change has removed row gap and now relies on margin but
it should be kept in sync with panel or any other container
the layout is being used.