Fix dummy fields dirty and restoration
There are two kinds of dummy fields we need to consider.

All dummy fields have in common the fact that they exist in the view, but not in the model.
"Clean" dummy fields are dummy fields that are prefixed by "$" or "__", and additionally cannot
cause the view to be dirty.

All dummy fields need to be kept after saving.
"Clean" dummy fields should not dirty the view.