Fix sync record in data store
When record is updated/read, it should update record in data store,
so that updated data will be available when switching back in grid view.

This make sure to update only necessary field in data store but also the
dot fields if present in the updated record. In case a field or dot field isn't
present in the form, if the value change during the save, updated value
will not be reflected in the grid view (but there is less chance that case happens).

Fixes #241

Signed-off-by: Pierre Belloy <p.belloy@axelor.com>