Fix dirty after canceling changes in editable grid and master-detail
While editing a row, dirty state is changed. If we cancel row edit,
need to recompute dirty state.

This recomputes dirty, instead of saving dirty state before edit and restoring it,
because in the case of master-detail, user is free to edit both details-view and parent fields,
causing what dirty state to go back to unclear in case of canceling.

Addresses #652