Reset grid offset on DMS search
As we can navigate through folders on the left side, the search
will keep the current offset. If a user goes to the next page
on a folder, he will see no records with a wrong text on the PageText
component if he clicks on another folder with fewer records.

Fixes #838

Signed-off-by: Pierre Belloy <p.belloy@axelor.com>