enable @typescript-eslint/no-shadow eslint rule
The rule disallow variable declarations from shadowing variables
declared in the outer scope. Shadowing is the process by which a local
variable shares the same name as a variable in its containing scope.

Shadowing can lead to unneeded bugs and results and will be harder to
debug.

See https://typescript-eslint.io/rules/no-shadow/ and
https://eslint.org/docs/latest/rules/no-shadow#hoist-all

Signed-off-by: Pierre Belloy <p.belloy@axelor.com>