Fix default value in deepGet utility
Suppose object is { x: null }, when we do deepGet({ x: null }, 'x'), it suppose to be return null for `x` value, but it return defaultValue i.e. undefined. So i have assigned current value as defaultValue.

In case of issue #1018, because of the binary value is undefined in record and on remove i used to reset it's value to null for additional processing at component level, but it still gets undefined value because of deepGet function which regression after 195bf7186761039e680a575b9647fa8e371acfbb commit.

Fixes #1018