Fix set custom fields attributes
Custom field that are part of the default `attrs` json field, attributes
can be updated either without prefix, no matter where the action is
triggered in the view. This means that whether the action is triggered
from a field event or a button in the main form or from a field event
or button inside a json field, it works same.

For custom fields that are part of other json fields, attributes have
to be updated with their respective prefix or if the action is executed
inside the json field, attributes can also be updated without prefix.

Fixes #1080