Add CacheBuilder.removalListener
Similar API to Caffeine.removalListener, to perform some operation when an entry is removed.

This is a common denominator between what's supported by both Caffeine and Redisson.

Most notably, there is no equivalent to Caffeine.evictionListener because Redisson lumps together
size eviction and explicit removals, so it's not possible to have unified behavior.

Addresses #1296