Remove useTemplate hook
Migrate useTemplate hook with TemplateRenderer component.

Reason of migration is with useTemplate hook it is sending Template component which is creating element on each time when context changes due to usage of React.createElement inside useMemo.

So through refactor we have separated the rendering of template and computation of context value.