Separate generateCode task for main and test sourceSet
Instead of a single task to generate main and test entities, split into
2 different tasks: the first one to generate only main entities
(aka `generateCode`) and a second to generate test entities
(aka `generateTestCode`). `compileTestJava` depends on `generateTestCode`.

As specified in the relationships between java tasks, compileTestJava
depends on classes, itself depending on compileJava. So, it will also
build main sourceSet so that they will be used for tests.

Signed-off-by: Pierre Belloy <p.belloy@axelor.com>