Rollback React Compiler support
The React compiler is still in RC. While the documentation indicates
it's safe to try it in production, no general availability has been
announced. While it's optional/experimental, it's safe to avoid using
it. A rollback will be hard as the new code gets rid of using `useMemo`,
`useCallback`, and `React.memo`.

Signed-off-by: Pierre Belloy <p.belloy@axelor.com>