Upgrade Postgresql JDBC from 42.7.4 to 42.7.8
In version >= 42.7.5, it comes with severe performance regression due
`DatabaseMetaData.getCrossReference(...)` method. This is commonly used
during startup at schema validation and generation: Hibernate inspects
the database schema to verify or generate mappings. For applications,
startup execution time was increased by 3x.

The fix is part of version 42.7.8, and the startup execution time is
now stable.

Signed-off-by: Pierre Belloy <p.belloy@axelor.com>