Disable mass update depending on entity inheritance strategy
Depending on inheritance strategy defined on entities, bulk update may
involve multiple tables. When he tries to update, Hibernate tries to
be smart and assigns "aliases" to manage the complexity. But PostgreSQL
strictly forbids these aliases in UPDATE statements and rejects the
query generated by Hibernate with a syntax error `missing FROM-clause`.

For that reason, mass update is now disabled depending on inheritance
strategy. This is the case for `CLASS` (ie, `TABLE_PER_CLASS`) and
`JOINED` strategy. For `CLASS`, as soon as it has a child class, it
is disabled. For `JOINED`, as soon as it has a subclass, it is disabled.

Fixes #1610

Signed-off-by: Pierre Belloy <p.belloy@axelor.com>